updates work, but then once the current drops it goes back to how it originally was (the problem state) and then alternates

+++ Team-3749-2024/src/main/java/frc/robot/utils/CurrentBudgettedSubsystem.java
@@ -1,6 +1,7 @@
 package frc.robot.utils;
 
 import java.util.function.DoubleSupplier;
+import java.util.function.IntSupplier;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Robot;
@@ -9,7 +10,7 @@ import frc.robot.subsystems.example.ExampleIO.ExampleData;
 public interface CurrentBudgettedSubsystem  {
 
     /** Updates the set of loggable inputs. */
-    public default void reduceCurrentSum(DoubleSupplier currentReductionSupplier) {
+    public default void reduceCurrentSum(IntSupplier currentReductionSupplier) {
 
     }
